# CHORE: Use faster test translation scenario, cut CI time by ~5mins