--- a/pyproject.toml
+++ b/pyproject.toml
@@ -55,6 +55,7 @@ test = [
   "tensorflow",
   "sentencepiece",
   "opencv-python",
+  "protobuf==3.20.0",
 ]
 
 [project.urls]