--- a/tests/explainers/conftest.py
+++ b/tests/explainers/conftest.py
@@ -11,8 +11,10 @@ def basic_translation_scenario():
     AutoTokenizer = pytest.importorskip("transformers").AutoTokenizer
     AutoModelForSeq2SeqLM = pytest.importorskip("transformers").AutoModelForSeq2SeqLM
 
-    tokenizer = AutoTokenizer.from_pretrained("Helsinki-NLP/opus-mt-en-es")
-    model = AutoModelForSeq2SeqLM.from_pretrained("Helsinki-NLP/opus-mt-en-es")
+    # Use a very small model, for speed
+    name = "mesolitica/finetune-translation-t5-super-super-tiny-standard-bahasa-cased"
+    tokenizer = AutoTokenizer.from_pretrained(name)
+    model = AutoModelForSeq2SeqLM.from_pretrained(name)
 
     # define the input sentences we want to translate
     data = [