--- a/tests/explainers/conftest.py
+++ b/tests/explainers/conftest.py
@@ -5,9 +5,10 @@ import pytest
 
 @pytest.mark.skipif(sys.platform == 'win32', reason="Integer division bug in HuggingFace on Windows")
 @pytest.fixture(scope="session")
-def basic_translation_scenario():
+def basic_translation_scenario(monkeypatch):
     """ Create a basic transformers translation model and tokenizer.
     """
+    monkeypatch.setenv("PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION", "python")
     AutoTokenizer = pytest.importorskip("transformers").AutoTokenizer
     AutoModelForSeq2SeqLM = pytest.importorskip("transformers").AutoModelForSeq2SeqLM